// src/types/enums/ordersInputTypes.ts
export enum EOrdersInputTypeKeys {
  Order = "order",
  Price = "price",
  DeliveryDayFrom = "deliveryDateFrom",
  DeliveryDayTo = "deliveryDateTo",
  Customer = "customer",
  DeliveryDay = "deliveryDate",
  Comment = "comment",
}

export enum EOrdersInputTypeTitles {
  Order = "Заказ",
  Price = "Цена",
  DeliveryDayFrom = "Дата доставки с",
  DeliveryDayTo = "Дата доставки до",
  Customer = "Заказчик",
  DeliveryDay = "Дата доставки",
  Comment = "Комментарий",
}